+++ demo/src/main/java/com/evernote/android/test/data/TestJavaGenericSerializable.java
@@ -1,0 +1,20 @@
+package com.evernote.android.test.data;
+
+import com.evernote.android.state.State;
+
+/**
+ * @author rwondratschek
+ */
+public class TestJavaGenericSerializable {
+
+    @State
+    private GenericSerializable<String> mGenericSerializable;
+
+    public GenericSerializable<String> getGenericSerializable() {
+        return mGenericSerializable;
+    }
+
+    public void setGenericSerializable(GenericSerializable<String> genericSerializable) {
+        mGenericSerializable = genericSerializable;
+    }
+}